refactor(ui): extract PersistenceMixin for file-dialogs + character load
Move the six character file / data-pack dialog methods out of L5RMain
into l5r/ui/persistence.py:

* load_character_from
* select_save_path
* select_load_path
* select_export_file
* select_import_data_pack
* create_new_character

Also drop the dead "supported_ext" locals in the two select_*
methods (assigned but never read).

main.py: 2300 -> 2170 lines.
