+++ l5r/main.py
@@ -56,11 +56,13 @@ from l5r.ui.helpers import (
     new_vert_line,
 )
 from l5r.ui.nicebar import NicebarMixin
+from l5r.ui.persistence import PersistenceMixin
 from l5r.util import log
 from l5r.util.settings import L5RCMSettings
 
 
-class L5RMain(AdviseMixin, HealthDisplayMixin, NicebarMixin, L5RCMCore):
+class L5RMain(AdviseMixin, HealthDisplayMixin, NicebarMixin,
+              PersistenceMixin, L5RCMCore):
 
     default_size = QtCore.QSize(820, 720)
     default_point_size = 8.25
@@ -1833,31 +1835,6 @@ class L5RMain(AdviseMixin, HealthDisplayMixin, NicebarMixin, L5RCMCore):
 
         self.update_from_model()
 
-    def load_character_from(self, path):
-
-        with QtSignalLock(self.pers_info_widgets + [self.tx_pc_name]):
-
-            if not self.pc:
-                self.create_new_character()
-
-            if self.pc.load_from(path):
-                self.save_path = path
-
-                if not api.character.books.fulfills_dependencies():
-                    # warn about missing dependencies
-                    self.warn_about_missing_books()
-
-                    # immediately create a new character
-                    self.create_new_character()
-                    return False
-
-                log.app.info('successfully loaded character from {0}'.format(self.save_path))
-
-                self.tx_pc_notes.set_content(self.pc.extra_notes)
-                self.update_from_model()
-            else:
-                log.app.error('character load failure')
-
     def set_clan(self, clan_id):
         """Set UI clan"""
 
@@ -2056,109 +2033,6 @@ class L5RMain(AdviseMixin, HealthDisplayMixin, NicebarMixin, L5RCMCore):
         else:
             super(L5RMain, self).closeEvent(ev)
 
-    def select_save_path(self):
-        settings = L5RCMSettings()
-        last_dir = settings.app.last_open_dir
-        char_name = self.get_character_full_name()
-        proposed = os.path.join(last_dir, char_name)
-
-        fileName = QtWidgets.QFileDialog.getSaveFileName(
-            self,
-            self.tr("Save Character"),
-            proposed,
-            self.tr("L5R Character files (*.l5r)"))
-
-        # user pressed cancel or didn't enter a name
-        if not fileName:
-            return None
-
-        # on pyqt5 it returns a tuple (fname, filter)
-        if type(fileName) is tuple:
-            fileName = fileName[0]
-
-        if fileName:
-            settings.app.last_open_dir = os.path.dirname(fileName)
-
-        if fileName.endswith('.l5r'):
-            return fileName
-        return fileName + '.l5r'
-
-    def select_load_path(self):
-        settings = L5RCMSettings()
-        last_dir = settings.app.last_open_dir
-        fileName = QtWidgets.QFileDialog.getOpenFileName(
-            self,
-            self.tr("Load Character"),
-            last_dir,
-            self.tr("L5R Character files (*.l5r)"))
-
-        # user pressed cancel or didn't enter a name
-        if not fileName:
-            return None
-
-        # on pyqt5 it returns a tuple (fname, filter)
-        if type(fileName) is tuple:
-            fileName = fileName[0]
-
-        if fileName:
-            settings.app.last_open_dir = os.path.dirname(fileName)
-        return fileName
-
-    def select_export_file(self, file_ext='.txt'):
-        supported_ext = ['.pdf']
-        supported_filters = [self.tr("PDF Files(*.pdf)")]
-
-        settings = L5RCMSettings()
-        last_dir = settings.app.last_open_dir
-        char_name = self.get_character_full_name()
-        proposed = os.path.join(last_dir, char_name)
-
-        fileName = QtWidgets.QFileDialog.getSaveFileName(
-            self,
-            self.tr("Export Character"),
-            proposed,
-            ";;".join(supported_filters))
-
-        # user pressed cancel or didn't enter a name
-        if not fileName:
-            return None
-
-        # on pyqt5 it returns a tuple (fname, filter)
-        if type(fileName) is tuple:
-            fileName = fileName[0]
-
-        if fileName:
-            settings.app.last_open_dir = os.path.dirname(fileName)
-
-        if fileName.endswith(file_ext):
-            return fileName
-        return fileName + file_ext
-
-    def select_import_data_pack(self):
-        supported_ext = ['.zip', '.l5rcmpack']
-        supported_filters = [self.tr("L5R:CM Data Pack(*.l5rcmpack *.zip)"),
-                             self.tr("Zip Archive(*.zip)")]
-
-        settings = L5RCMSettings()
-        last_data_dir = settings.app.last_open_data_dir
-
-        files = QtWidgets.QFileDialog.getOpenFileNames(
-            self,
-            self.tr("Load data pack"),
-            last_data_dir,
-            ";;".join(supported_filters))
-
-        if type(files) is tuple:
-            files = files[0]
-
-        if not files:
-            return None
-
-        if files[0]:
-            settings.app.last_open_data_dir = os.path.dirname(files[0])
-
-        return files
-
     def on_change_insight_calculation(self):
         method = self.sender().checkedAction().property('method')
         api.character.set_insight_calculation_method(method)
@@ -2169,10 +2043,6 @@ class L5RMain(AdviseMixin, HealthDisplayMixin, NicebarMixin, L5RCMCore):
         settings = L5RCMSettings()
         settings.app.health_method = method
         self.update_from_model()
-
-    def create_new_character(self):
-        self.sink1.new_character()
-        self.pc.unsaved = False
 
 # MAIN ###
